requirements.txt
google-genai>=2.0.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tqdm>=4.66.0
requests>=2.31.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
